Remove sales tax display and share button
This commit removes the conditional block for displaying the sales tax amount in `CreatePaymentRequest.cshtml` and eliminates the button for sharing the QR code. The print button and new payment request link remain intact. Additionally, in `PaymentSuccess.cshtml`, the link for downloading the transaction invoice has been removed, while the options for performing another payment and returning to the user panel are preserved.